Combine metrics classes into one
