--- src/Agent.Worker/ResourceMetrics/IResourceDataCollector.cs
@@ -1,7 +1,0 @@
-﻿namespace Microsoft.VisualStudio.Services.Agent.Worker.ResourceMetrics
-{
-    public interface IResourceDataCollector
-    {
-        string GetCurrentData(ITerminal terminal);
-    }
-}
